feat: refactor mobile banking items
